requirements.txt
streamlit
openai-whisper
torch
indic-transliteration